services/langchain/requirements.txt
beautifulsoup4==4.12.2
chromadb==0.4.15
langchain==0.0.329
faiss_cpu==1.7.4
openai==0.28.1
pypdf==3.17.0
PyPDF2==3.0.1
pysqlite3-binary==0.5.2.post1
python-dotenv==1.0.0
streamlit==1.28.0
tiktoken==0.5.1
